+++ components/GuessDisplay.tsx
