+++ components/GuessDisplayH.tsx
